perf: 构建队列MQ的弹性最大并发线程保持一致 #2994 增加耗时的监控日志
