+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/listener/pipeline/MQPipelineCreateListener.kt
@@ -29,6 +29,7 @@ package com.tencent.devops.process.engine.listener.pipeline
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.tencent.devops.common.api.enums.RepositoryTypeNew
 import com.tencent.devops.common.api.enums.ScmType
+import com.tencent.devops.common.api.util.Watcher
 import com.tencent.devops.common.event.dispatcher.pipeline.PipelineEventDispatcher
 import com.tencent.devops.common.event.listener.pipeline.BaseListener
 import com.tencent.devops.common.pipeline.pojo.element.Element
@@ -41,6 +42,7 @@ import com.tencent.devops.common.pipeline.pojo.element.trigger.CodeTGitWebHookTr
 import com.tencent.devops.common.pipeline.pojo.element.trigger.enums.CodeEventType
 import com.tencent.devops.common.pipeline.utils.RepositoryConfigUtils
 import com.tencent.devops.common.redis.RedisOperation
+import com.tencent.devops.common.service.utils.LogUtils
 import com.tencent.devops.common.websocket.constant.WebsocketCode
 import com.tencent.devops.common.websocket.dispatch.WebSocketDispatcher
 import com.tencent.devops.common.websocket.enum.NotityLevel
@@ -56,7 +58,6 @@ import com.tencent.devops.process.websocket.page.EditPageBuild
 import com.tencent.devops.process.websocket.push.WebHookWebsocketPush
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
-import org.springframework.util.StopWatch
 
 /**
  *  MQ实现的流水线创建事件
@@ -74,33 +75,36 @@ class MQPipelineCreateListener @Autowired constructor(
 ) : BaseListener<PipelineCreateEvent>(pipelineEventDispatcher) {
 
     override fun run(event: PipelineCreateEvent) {
-        val watch = StopWatch("pipelineCreateEventWatch")
-        val pipelineId = event.pipelineId
-        if (event.source == ("create_pipeline")) {
-            watch.start("createPipeline[pipelineId:$pipelineId]")
-            callBackControl.pipelineCreateEvent(projectId = event.projectId, pipelineId = pipelineId)
-            watch.stop()
+        val watcher = Watcher(id = "${event.traceId}_CreatePipeline#${event.pipelineId}_${event.userId}")
+        try {
+
+            if (event.source == ("create_pipeline")) {
+                watcher.start("callback")
+                callBackControl.pipelineCreateEvent(projectId = event.projectId, pipelineId = event.pipelineId)
+                watcher.stop()
+            }
+            if (event.source == "createWebhook") {
+                watcher.start("createWebHook")
+                addWebHook(event.element!!, event)
+                watcher.stop()
+            }
+        } finally {
+            watcher.stop()
+            LogUtils.printCostTimeWE(watcher = watcher)
         }
-        if (event.source == "createWebhook") {
-            logger.info("[$pipelineId] createGitWebhook!MQ内调用")
-            watch.start("createWebHook[pipelineId:$pipelineId]")
-            addWebHook(event.element!!, event)
-            watch.stop()
-        }
-        logger.info("pipelineId:$pipelineId pipelineCreateEvent watch is:$watch")
     }
 
     private fun addWebHook(e: Element, event: PipelineCreateEvent) {
         val (repositoryConfig, scmType, eventType) = when (e) {
             is CodeGitWebHookTriggerElement -> Triple(
-                    RepositoryConfigUtils.buildConfig(e),
-                    ScmType.CODE_GIT,
-                    e.eventType
+                RepositoryConfigUtils.buildConfig(e),
+                ScmType.CODE_GIT,
+                e.eventType
             )
             is CodeGitlabWebHookTriggerElement -> Triple(
-                    RepositoryConfigUtils.buildConfig(e),
-                    ScmType.CODE_GITLAB,
-                    null
+                RepositoryConfigUtils.buildConfig(e),
+                ScmType.CODE_GITLAB,
+                null
             )
             is CodeSVNWebHookTriggerElement -> Triple(RepositoryConfigUtils.buildConfig(e), ScmType.CODE_SVN, null)
             is CodeGithubWebHookTriggerElement -> Triple(RepositoryConfigUtils.buildConfig(e), ScmType.GITHUB, null)
@@ -175,20 +179,20 @@ class MQPipelineCreateListener @Autowired constructor(
                 }
             } catch (e: Exception) {
                 val post = NotifyPost(
-                        module = "process",
-                        message = e.message!!,
-                        level = NotityLevel.HIGH_LEVEL.getLevel(),
-                        dealUrl = EditPageBuild().buildPage(
-                                BuildPageInfo(
-                                        projectId = event.projectId,
-                                        pipelineId = event.pipelineId,
-                                        buildId = null,
-                                        atomId = null
-                                )
-                        ),
-                        code = WebsocketCode.WEBHOOK_ADD_ERROR,
-                        webSocketType = WebSocketType.changWebType(WebSocketType.WEBHOOK),
-                        page = null
+                    module = "process",
+                    message = e.message!!,
+                    level = NotityLevel.HIGH_LEVEL.getLevel(),
+                    dealUrl = EditPageBuild().buildPage(
+                        BuildPageInfo(
+                            projectId = event.projectId,
+                            pipelineId = event.pipelineId,
+                            buildId = null,
+                            atomId = null
+                        )
+                    ),
+                    code = WebsocketCode.WEBHOOK_ADD_ERROR,
+                    webSocketType = WebSocketType.changWebType(WebSocketType.WEBHOOK),
+                    page = null
                 )
                 websocketDispatch(post, event)
                 logger.error("[${event.pipelineId}]异步调用webhook返回未知异常。webSocket推送异常信息[$post]", e)
@@ -198,17 +202,17 @@ class MQPipelineCreateListener @Autowired constructor(
 
     private fun websocketDispatch(notifyPost: NotifyPost, event: PipelineCreateEvent) {
         webSocketDispatcher.dispatch(
-                WebHookWebsocketPush(
-                        buildId = null,
-                        pipelineId = event.pipelineId,
-                        projectId = event.projectId,
-                        userId = event.userId,
-                        pushType = WebSocketType.WEBHOOK,
-                        redisOperation = redisOperation,
-                        page = notifyPost.dealUrl,
-                        objectMapper = objectMapper,
-                        notifyPost = notifyPost
-                )
+            WebHookWebsocketPush(
+                buildId = null,
+                pipelineId = event.pipelineId,
+                projectId = event.projectId,
+                userId = event.userId,
+                pushType = WebSocketType.WEBHOOK,
+                redisOperation = redisOperation,
+                page = notifyPost.dealUrl,
+                objectMapper = objectMapper,
+                notifyPost = notifyPost
+            )
         )
     }
 }
