+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/listener/pipeline/MQPipelineUpdateListener.kt
@@ -26,11 +26,12 @@
 
 package com.tencent.devops.process.engine.listener.pipeline
 
+import com.tencent.devops.common.api.util.Watcher
 import com.tencent.devops.common.event.dispatcher.pipeline.PipelineEventDispatcher
 import com.tencent.devops.common.event.listener.pipeline.BaseListener
+import com.tencent.devops.common.service.utils.LogUtils
 import com.tencent.devops.process.engine.control.CallBackControl
 import com.tencent.devops.process.engine.pojo.event.PipelineUpdateEvent
-import com.tencent.devops.process.engine.service.PipelineRepositoryService
 import com.tencent.devops.process.engine.service.PipelineRuntimeService
 import com.tencent.devops.process.service.PipelineUserService
 import org.springframework.beans.factory.annotation.Autowired
@@ -45,23 +46,25 @@ import org.springframework.stereotype.Component
 class MQPipelineUpdateListener @Autowired constructor(
     private val pipelineUserService: PipelineUserService,
     private val pipelineRuntimeService: PipelineRuntimeService,
-    private val pipelineRepositoryService: PipelineRepositoryService,
     private val callBackControl: CallBackControl,
     pipelineEventDispatcher: PipelineEventDispatcher
 ) : BaseListener<PipelineUpdateEvent>(pipelineEventDispatcher) {
 
     override fun run(event: PipelineUpdateEvent) {
-        if (event.buildNo != null) {
-            pipelineRuntimeService.updateBuildNo(event.pipelineId, event.buildNo.buildNo)
-            logger.info("[${event.pipelineId}] updateBuildNo!")
+        val watcher = Watcher(id = "${event.traceId}_UpdatePipeline#${event.pipelineId}_${event.userId}")
+        try {
+            if (event.buildNo != null) {
+                watcher.start("updateBuildNo")
+                pipelineRuntimeService.updateBuildNo(event.pipelineId, event.buildNo.buildNo)
+                watcher.stop()
+            }
+
+            pipelineUserService.update(event.pipelineId, event.userId)
+            watcher.start("callback")
+            callBackControl.pipelineUpdateEvent(projectId = event.projectId, pipelineId = event.pipelineId)
+        } finally {
+            watcher.stop()
+            LogUtils.printCostTimeWE(watcher)
         }
-        val model = pipelineRepositoryService.getModel(event.pipelineId)
-        if (model == null || model.stages.isEmpty()) {
-            logger.warn("[${event.pipelineId}]|pipeline model stage is empty")
-            return
-        }
-        val pipelineId = event.pipelineId
-        pipelineUserService.update(pipelineId, event.userId)
-        callBackControl.pipelineUpdateEvent(projectId = event.projectId, pipelineId = event.pipelineId)
     }
 }
